初步实现编译

+++ fascript/fascript.hpp
@@ -14,8 +14,8 @@ class IAstExpr;
 class FAScript: std::enable_shared_from_this<FAScript> {
 public:
 	Value RunCode (std::string _code);
-	uint16_t GetNameId (AstIdType _type, std::string _name);
-	uint16_t NewGlobalFuncId (std::shared_ptr<Function> _func);
+	int32_t GetNameId (AstIdType _type, std::string _name);
+	int32_t NewGlobalFuncId (std::shared_ptr<Function> _func);
 
 	// ×¢²á´ý±àÒëº¯Êý
 	void RegisterUncompiledFunc (std::shared_ptr<IAstExpr> _ast_func) { m_uncompiled_funcs.push_back (_ast_func); }
@@ -27,13 +27,14 @@ public:
 
 private:
 	std::recursive_mutex m_mtx;
-	std::map<std::string, uint16_t> m_name_to_id;
-	std::map<uint16_t, Value> m_id_to_var;
-	uint16_t m_next_id = 0;
+	std::map<std::string, int32_t> m_name_to_id;
+	std::map<int32_t, Value> m_id_to_var;
+
+	int32_t m_next_id = 0;
 
 	std::vector<std::shared_ptr<IAstExpr>> m_uncompiled_funcs;
-	std::map<uint16_t, std::shared_ptr<Function>> m_id_to_func;
-	uint16_t m_next_func_id = 0;
+	std::map<int32_t, std::shared_ptr<Function>> m_id_to_func;
+	int32_t m_next_func_id = 0;
 
 	std::vector<std::shared_ptr<IAstExpr>> m_current_blocks;
 };
